refactor: organize service test files into a single directory
- Create a `tests` directory under `src/services/`
- Move `TaskManager.test.ts` into the `tests` directory and update import paths
- Combine `TaskSanitizer.test.ts` and `__tests__/TaskSanitizer.test.ts` into a single file in the `tests` directory and update import paths
- Remove old test files and the `__tests__` directory
